requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pyyaml==6.0.1
uritemplate==4.1.1
inflection==0.5.1